+++ lib/app/app_widget.dart
@@ -1,0 +1,37 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_2/app/pages/home_page/home.dart';
+
+class MyApp extends StatefulWidget {
+  @override
+  _MyAppState createState() => _MyAppState();
+}
+
+class _MyAppState extends State<MyApp> {
+  double currentSize = 0;
+
+  @override
+  Widget build(BuildContext context) {
+    return MaterialApp(
+      debugShowCheckedModeBanner: false,
+      title: 'Material App',
+      home: Scaffold(
+        appBar: AppBar(
+          backgroundColor: Colors.red[800], //Color(0xFF3c8dbc),
+          elevation: 0,
+          title: Text(
+            'Flutter Dashboard',
+            style: TextStyle(
+              fontSize: 15,
+            ),
+          ),
+          actions: [
+            IconButton(icon: Icon(Icons.mail_outline), onPressed: () {})
+          ],
+        ),
+        drawer: Drawer(),
+        drawerEdgeDragWidth: 200,
+        body: Home(),
+      ),
+    );
+  }
+}
