+++ lib/app/models/user.dart
