+++ lib/app/tabs/usersTab.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_2/models/user.dart';
+import 'package:flutter_2/app/models/user.dart';
 
 class UserTab extends StatefulWidget {
   const UserTab({Key key}) : super(key: key);
